refactor: ImageUpload 컴포넌트 공통 컴포넌트로 분리 (#20)
