+++ app/project/_components/index.ts
@@ -1,4 +1,3 @@
-export { default as ImageUpload } from './image-upload'
 export { default as Title } from './title'
 export { default as Overview } from './overview'
 export { default as Site } from './site'
