+++ app/project/_models/index.ts
@@ -4,4 +4,3 @@ export * from './members'
 export * from './stack'
 export * from './roles'
 export * from './feature'
-export * from './image'
