+++ types/index.ts
@@ -1,2 +1,3 @@
 export * from './common'
 export * from './card'
+export * from './image'
